made useTImer hook and convertTime resuable function
