+++ src/components/forms/auth/ResetPasswordCodeFrom.tsx
@@ -22,7 +22,6 @@ const ResetPasswordCodeForm = ({
   const {
     register,
     handleSubmit,
-    watch,
     formState: { errors },
   } = useForm<Code>();
